#28 UIButton 관련 설정을 extension으로 분리
